move inline TemplateEditor
