+++ panel/src/components/Commands/common/templates/TemplateEditor.css
